examples/deployment/flyio-example/requirements.txt
pipecat-ai[daily,openai,silero]
fastapi
uvicorn
requests
python-dotenv
loguru